requirements.txt
speechrecognition==3.10.0
pyttsx3==2.90
pywhatkit==5.4
wikipedia==1.4.0
pyjokes==0.6.0
python-dotenv==1.0.0
requests==2.31.0
playsound==1.3.0
datetime==4.3
